Move: VMStatusController component
